requirements.txt
matplotlib
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
python-dotenv==1.0.1
requests
scanpy==1.10.2
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
streamlit==1.37.1
gdown==5.2.0
